chore: renamed openapi file to handler
