+++ src/index.ts
@@ -1,4 +1,4 @@
-export { generateSpecs, openAPISpecs } from "./openapi.js";
+export { generateSpecs, openAPISpecs } from "./handler.js";
 export type * from "./types.js";
 export * from "./utils.js";
 export * from "./middlewares.js";
